Strip userid field from response to /token.
The user-ID is just an incrementing integer from the database. To
deliver that as identifier to the outside weakens this bot
unnecessarily: If the incrementation is resetted (e.g.  database
crashes, is not copied on updates, ...), then the IDs suddenly reference
other contacts.
Exposing the email address as identifier to the outside avoids that.